enable DEV_NAME to handle symlinks
Using a macro is not possible anymore, so have a function dev_name() do
the job. Functionality is the same as the old DEV_NAME, unless the
specified path is a symlink in which case it's being dereferenced first.

To save callers from having to free the returned string, a static local
buffer is used (effectively rendering the function non-reentrant).