Fix comments in TEXT section
When a '#' is found in TEXT, section remove_comments() is called.  However, that function doesn't
limit itself to removing that particular comment. Instead, it greedily removes all "comments" from
the string, including those in ${color $abcdef} constructs. This, obviously, makes a mess of
things.

Solution: I added a new function, remove_comment(), that only removes the current comment.

Note: Atm, the function doesn't delete the newline, since that was the current behavior, and is
the natural one for the config section. However, for the TEXT it seems better to delete the \n
too, to avoid blank lines.

Note2: this also fixes the bug in strfold(), which incorectly copied the string, truncating it by
one byte.

Signed-off-by: Brenden Matthews <brenden@rty.ca>