Added own_window_type panel
The panel type reserves space along the edge of the
screen, just like regular DE panels, taskbars and the
like. I have tested it for several hours now with lots
of different settings, and it seems to be working fine.

Works especially well with alignment top_??? or
bottom_??? and single-line output. Something like
this:

  own_window yes
  own_window_type panel

  alignment bottom_left
  maximum_width 1680
  minimum_size 1680
  gap_x 0
  gap_y 0

  stippled_borders 1
  draw_borders yes

  update_interval 3.0

  TEXT
  $nodename $tab $freq_g ${color grey}GHz$color $tab $memperc% ${color grey}RAM$color $tab $cpu% ${color grey}CPU $tab etc...etc...$alignr${time %F %R}

The line after TEXT is supposed to be on one line,
just in case some channel has messed it up.

Signed-off-by: Brenden Matthews <brenden@rty.ca>