Read MPD connection data from environment.
Most MPD clients read the MPD_HOST and MPD_PORT
environment variables. Now, conky will too.

MPD_HOST can be either "hostname" or "password@hostname".

If a user specifies a host in the configuration, the
password set in MPD_HOST will be ignored. This is to
prevent the password from being sent to the wrong host.

In other words, if the host is specified in the conky
configuration, the password must be too (if there is
one).

Signed-off-by: Brenden Matthews <brenden@rty.ca>