Fix: conky waited indefinitely on programs started with \$exec-alike vars if those programs don't quit.
Conky now kills the program when it should start the next update, this makes using things like tail with
the -f option possible in a \$exec.

I am not pushing this to 1.7.2 because this is a pretty big change in the code and it is not really
a bugfix but more a usability-problem-fix (if that term would exist).