Hide false memleaks in a debugbuild
This makes sure valgrind doesn't show false memleaks that are caused by ncurses
instead of our code when we do a 'Debug' build. There is a also a disadvantage:
When your ncurses library is not build with the '--disable-leaks' option (this
is the case in the default version of ncurses of most distro's) and you build
conky with ncurses and debug turned on then you will see a segmentation fault
when you stop conky. On the other hand people that want to build conky as a
debug build with ncurses support will probably also have a ncurses with
--disable-leaks ...