Support imap passwords with strange characters (sf.net #3402448)
conky now sends the password using the length + string format, which should not need any quoting
of special characters. Hopefully, we won't find a server which does not support this encoding.

This also fixes a bug which prevented the password from being entered from the console.

bug reported by Brent