Fix a crash on exit when using curl
The problem was that the callback thread is destroyed only in the destructor of the callback_base
class (which is called after the destructor of the derived classes). This means that the thread
is running even when it's object is partly destroyed, which can cause segfaults, race conditions
and other nasty problems.

I've fixed it so that the thread is destroyed before the underlying object's destructor is
called.