Add configuration setting imlib_image_alpha
This option is passed verbatim to both calls to
imlib_image_set_has_alpha. When false, imlib will not merge the alpha
channel when blending the image into the window visual, which means that
own_window_argb_visual will not result in transparent images. The option
is enabled by default to preserve the current behaviour.