linux: Fix possible long overflow
The C language defines long to be at least 32 bit and so on ILP32 data
models like Win32 and 32-bit Unix systems. It therefore will overflow
for values of the order of 1e12, which we obtain as a multiplication
result in get_battery_power_draw().

Fix this by an early conversion to double.