Use WithdrawnState only for Fluxbox
Openbox completely ignores all atoms set for Conky window if it's started
in WithdrawnState. Fluxbox requires docks to be started like this.

Documentation doesn't really state when WithdrawnState should be used,
but in general it means "window isn't managed by WM" (based on CPAN docs).
We want normal windows (not override) to be managed by the WM, so
NormalState is probably a better default for dock and panel windows.

Signed-off-by: Tin Švagelj <tin.svagelj@live.com>