Use WithdrawnState only for Fluxbox (#2196)
Openbox completely ignores all atoms set for Conky window if it's started in WithdrawnState. Fluxbox requires docks to be started like this.

Documentation doesn't really state when WithdrawnState should be used, but in general it means "window isn't managed by WM and not shown/mapped" (based on CPAN docs). We want normal windows (not override) to be managed by the WM, so NormalState is probably a better default for dock and panel windows because they are managed by WMs and shown when created.

This means that Fluxbox users will have to set XDG_CURRENT_DESKTOP to Fluxbox manually, like some other apps require them to do as well (see mailing list), in order to get the expected behavior.

Signed-off-by: Tin Švagelj <tin.svagelj@live.com>