Fix background rendering for wayland output
Previously, the background was rendered with cairo's CLEAR operator
which just clears the surface and does not render any visible contents.

The changes here first clear the surface (with CAIRO_OPERATOR_CLEAR) and
then use (CAIRO_OPERATOR_OVER) to render the contents. The result is
that the background colour is rendered correctly, including with
semi-transparent colours.